:boar:
This will probably work on the iPad.  I'm getting this error:
Exception '[<NSError 0x7f884add8290> setValue:forUndefinedKey:]: this class is not key value coding-compliant for the key name.'

Pretty sure that's from the [err setValue:@"File does not exist" forKey:@"description"]; exception handler in Firestack.m

The simulator doesn't actually support a camera, so it might not save to a fake location.